myproject/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.7
django-bootstrap3==15.0.0
djangorestframework==3.12.4
pyjsparser==2.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
tzlocal==3.0
